backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
pydantic[email]==2.5.0
python-multipart==0.0.6

